rename analytical pulses module since it is now available for all algorithms
