+++ src/qutip_qoc/global_local_opt.py
